+++ lib/cgraph/refstr.c
@@ -13,60 +13,293 @@
  * Contributors: Details at https://graphviz.org
  *************************************************************************/
 
+#include <assert.h>
 #include <cgraph/cghdr.h>
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 #include <util/alloc.h>
+#include <util/unreachable.h>
 
 /*
  * reference counted strings.
  */
 
 typedef struct {
-    Dtlink_t link;
     uint64_t refcnt: sizeof(uint64_t) * 8 - 1;
     uint64_t is_html: 1;
     char *s;
     char store[1];		/* this is actually a dynamic array */
 } refstr_t;
 
-static Dtdisc_t Refstrdisc = {
-    offsetof(refstr_t, s),	/* key */
-    -1,				/* size */
-    0,				/* link offset */
-    NULL,
-    free,
-    NULL,
-};
+/// compare reference-counted strings for equality
+static bool refstr_eq(const refstr_t *a, const refstr_t *b) {
+  return strcmp(a->s, b->s) == 0;
+}
 
-static Dict_t *Refdict_default;
+/// a string dictionary
+typedef struct {
+  refstr_t **buckets;  ///< backing store of elements
+  size_t size;         ///< number of elements in the dictionary
+  size_t capacity_exp; ///< log₂ size of `buckets`
+} strdict_t;
+
+static strdict_t *Refdict_default;
+
+/// derive a hash value from the given data
+///
+/// @param key Start of data to read
+/// @param len Number of bytes to read
+/// @return A hash digest suitable for dictionary indexing
+static uint64_t hash(const void *key, size_t len) {
+  assert(key != NULL || len == 0);
+
+  // The following implementation is based on the `MurmurHash64A` variant of the
+  // public domain MurmurHash by Austin Appleby. More information on this at
+  // https://github.com/aappleby/smhasher/. Relevant changes made to Austin’s
+  // original implementation:
+  //   • Our implementation is alignment-agnostic. No assumption is made about
+  //     the initial alignment of `key`.
+  //   • Our implementation uses `unsigned char` pointers, avoiding Undefined
+  //     Behavior when the input pointer originated from a non-`uint64_t`
+  //     object. This is written in a style that allows contemporary compilers
+  //     to optimize code back into wider 8-byte accesses where possible.
+
+  static const uint64_t seed = 0;
+
+  const uint64_t m = UINT64_C(0xc6a4a7935bd1e995);
+  const unsigned r = 47;
+
+  uint64_t h = seed ^ (len * m);
+
+  const unsigned char *data = key;
+  const unsigned char *end = data + len / sizeof(uint64_t) * sizeof(uint64_t);
+
+  while (data != end) {
+
+    uint64_t k;
+    memcpy(&k, data, sizeof(k));
+    data += sizeof(k);
+
+    k *= m;
+    k ^= k >> r;
+    k *= m;
+
+    h ^= k;
+    h *= m;
+  }
+
+  const unsigned char *data2 = data;
+
+  switch (len & 7) {
+  case 7:
+    h ^= (uint64_t)data2[6] << 48; // fall through
+  case 6:
+    h ^= (uint64_t)data2[5] << 40; // fall through
+  case 5:
+    h ^= (uint64_t)data2[4] << 32; // fall through
+  case 4:
+    h ^= (uint64_t)data2[3] << 24; // fall through
+  case 3:
+    h ^= (uint64_t)data2[2] << 16; // fall through
+  case 2:
+    h ^= (uint64_t)data2[1] << 8; // fall through
+  case 1:
+    h ^= (uint64_t)data2[0];
+    h *= m;
+    break;
+  default:
+    // nothing required
+    break;
+  }
+
+  h ^= h >> r;
+  h *= m;
+  h ^= h >> r;
+
+  return h;
+}
+
+/// create a new string dictionary
+static strdict_t *strdict_new(void) { return gv_alloc(sizeof(strdict_t)); }
+
+/// derive hash for a given reference-counted string
+static size_t strdict_hash(const refstr_t *key) {
+  assert(key != NULL);
+  return (size_t)hash(key->s, strlen(key->s));
+}
+
+/// a sentinel, marking a dictionary bucket from which an element has been
+/// deleted
+static refstr_t *const TOMBSTONE = (refstr_t *)-1;
+
+/// add a reference-counted string to a dictionary
+static void strdict_add(strdict_t *dict, refstr_t *r) {
+  assert(dict != NULL);
+  assert(r != NULL);
+  assert(r != TOMBSTONE);
+
+  // a watermark ratio at which the set capacity should be expanded
+  static const size_t OCCUPANCY_THRESHOLD_PERCENT = 70;
+
+  // do we need to expand the backing store?
+  size_t capacity = dict->buckets == NULL ? 0 : 1ul << dict->capacity_exp;
+  const bool grow = 100 * dict->size >= OCCUPANCY_THRESHOLD_PERCENT * capacity;
+
+  if (grow) {
+    const size_t new_c = capacity == 0 ? 10 : dict->capacity_exp + 1;
+    refstr_t **new_b = gv_calloc(1ul << new_c, sizeof(refstr_t *));
+
+    // Construct a new dictionary and copy everything into it. Note we need to
+    // rehash because capacity (and hence modulo wraparound behavior) has
+    // changed. This conveniently flushes out the tombstones too.
+    strdict_t new_d = {.buckets = new_b, .capacity_exp = new_c};
+    for (size_t i = 0; i < capacity; ++i) {
+      // skip empty buckets
+      if (dict->buckets[i] == NULL) {
+        continue;
+      }
+      // skip deleted buckets
+      if (dict->buckets[i] == TOMBSTONE) {
+        continue;
+      }
+      strdict_add(&new_d, dict->buckets[i]);
+    }
+
+    // replace ourselves with this new dictionary
+    free(dict->buckets);
+    *dict = new_d;
+  }
+
+  assert(dict->buckets != NULL);
+  capacity = 1ul << dict->capacity_exp;
+  assert(capacity > dict->size);
+
+  const size_t h = strdict_hash(r);
+
+  for (size_t i = 0; i < capacity; ++i) {
+    const size_t candidate = (h + i) % capacity;
+
+    // if we found an empty bucket or a previously deleted bucket, we can insert
+    if (dict->buckets[candidate] == NULL ||
+        dict->buckets[candidate] == TOMBSTONE) {
+      dict->buckets[candidate] = r;
+      ++dict->size;
+      return;
+    }
+  }
+
+  UNREACHABLE();
+}
+
+/// lookup a reference-counted string in a dictionary
+static refstr_t *strdict_find(strdict_t *dict, refstr_t *key) {
+  assert(dict != NULL);
+  assert(key != NULL);
+  assert(key != TOMBSTONE);
+
+  const size_t h = strdict_hash(key);
+  const size_t capacity = dict->buckets == NULL ? 0 : 1ul << dict->capacity_exp;
+
+  for (size_t i = 0; i < capacity; ++i) {
+    const size_t candidate = (h + i) % capacity;
+
+    // if we found an empty bucket, the sought item does not exist
+    if (dict->buckets[candidate] == NULL) {
+      return NULL;
+    }
+
+    // if we found a previously deleted slot, skip over it
+    if (dict->buckets[candidate] == TOMBSTONE) {
+      continue;
+    }
+
+    // is this the string we are searching for?
+    if (refstr_eq(key, dict->buckets[candidate])) {
+      return dict->buckets[candidate];
+    }
+  }
+
+  // not found
+  return NULL;
+}
+
+/// remove a reference-counted string from a dictionary
+static void strdict_remove(strdict_t *dict, const refstr_t *key) {
+  assert(dict != NULL);
+  assert(key != NULL);
+  assert(key != TOMBSTONE);
+
+  const size_t h = strdict_hash(key);
+  const size_t capacity = dict->buckets == NULL ? 0 : 1ul << dict->capacity_exp;
+
+  for (size_t i = 0; i < capacity; ++i) {
+    const size_t candidate = (h + i) % capacity;
+
+    // if we found an empty bucket, the sought item does not exist
+    if (dict->buckets[candidate] == NULL) {
+      return;
+    }
+
+    // if we found a previously deleted bucket, skip over it
+    if (dict->buckets[candidate] == TOMBSTONE) {
+      continue;
+    }
+
+    // is this the string we are searching for?
+    if (refstr_eq(key, dict->buckets[candidate])) {
+      assert(dict->size > 0);
+      free(dict->buckets[candidate]);
+      dict->buckets[candidate] = TOMBSTONE;
+      --dict->size;
+      return;
+    }
+  }
+}
+
+/// destroy a string dictionary
+static void strdict_free(strdict_t **dict) {
+  assert(dict != NULL);
+
+  if (*dict != NULL && (*dict)->buckets != NULL) {
+    for (size_t i = 0; i < 1ul << (*dict)->capacity_exp; ++i) {
+      if ((*dict)->buckets[i] != TOMBSTONE) {
+        free((*dict)->buckets[i]);
+      }
+    }
+    free((*dict)->buckets);
+  }
+
+  free(*dict);
+  *dict = NULL;
+}
 
 /* refdict:
- * Return the string dictionary associated with g.
+ * Return a pointer to the string dictionary associated with g.
  * If necessary, create it.
  */
-static Dict_t *refdict(Agraph_t * g)
-{
-    Dict_t **dictref;
+static strdict_t **refdict(Agraph_t *g) {
+    strdict_t **dictref;
 
     if (g)
-	dictref = &(g->clos->strdict);
+	dictref = (strdict_t **)&g->clos->strdict;
     else
 	dictref = &Refdict_default;
     if (*dictref == NULL) {
-	*dictref = agdtopen(g, &Refstrdisc, Dttree);
+	*dictref = strdict_new();
     }
-    return *dictref;
+    return dictref;
 }
 
 int agstrclose(Agraph_t * g)
 {
-    return agdtclose(g, refdict(g));
+    strdict_free(refdict(g));
+    return 0;
 }
 
-static refstr_t *refsymbind(Dict_t * strdict, const char *s)
-{
+static refstr_t *refsymbind(strdict_t *strdict, const char *s) {
     refstr_t key, *r;
 // Suppress Clang/GCC -Wcast-qual warning. Casting away const here is acceptable
 // as dtsearch does not modify its input key.
@@ -78,12 +311,11 @@ static refstr_t *refsymbind(Dict_t * strdict, const char *s)
 #ifdef __GNUC__
 #pragma GCC diagnostic pop
 #endif
-    r = dtsearch(strdict, &key);
+    r = strdict_find(strdict, &key);
     return r;
 }
 
-static char *refstrbind(Dict_t * strdict, const char *s)
-{
+static char *refstrbind(strdict_t *strdict, const char *s) {
     refstr_t *r;
     r = refsymbind(strdict, s);
     if (r)
@@ -94,17 +326,16 @@ static char *refstrbind(Dict_t * strdict, const char *s)
 
 char *agstrbind(Agraph_t * g, const char *s)
 {
-    return refstrbind(refdict(g), s);
+    return refstrbind(*refdict(g), s);
 }
 
 static char *agstrdup_internal(Agraph_t *g, const char *s, bool is_html) {
     refstr_t *r;
-    Dict_t *strdict;
     size_t sz;
 
     if (s == NULL)
 	 return NULL;
-    strdict = refdict(g);
+    strdict_t *strdict = *refdict(g);
     r = refsymbind(strdict, s);
     if (r)
 	r->refcnt++;
@@ -122,7 +353,7 @@ static char *agstrdup_internal(Agraph_t *g, const char *s, bool is_html) {
 	r->is_html = is_html;
 	strcpy(r->store, s);
 	r->s = r->store;
-	dtinsert(strdict, r);
+	strdict_add(strdict, r);
     }
     return r->s;
 }
@@ -138,17 +369,16 @@ char *agstrdup_html(Agraph_t *g, const char *s) {
 int agstrfree(Agraph_t * g, const char *s)
 {
     refstr_t *r;
-    Dict_t *strdict;
 
     if (s == NULL)
 	 return FAILURE;
 
-    strdict = refdict(g);
+    strdict_t *strdict = *refdict(g);
     r = refsymbind(strdict, s);
     if (r && r->s == s) {
 	r->refcnt--;
 	if (r->refcnt == 0) {
-	    agdtdelete(g, strdict, r);
+	    strdict_remove(strdict, r);
 	}
     }
     if (r == NULL)
@@ -190,6 +420,10 @@ static int refstrprint(void *ptr, void *user) {
 
 void agrefstrdump(Agraph_t * g)
 {
-    dtwalk(refdict(g), refstrprint, 0);
+    const strdict_t *d = *refdict(g);
+    for (size_t i = 0;
+         d != NULL && d->buckets != NULL && i < 1ul << d->capacity_exp; ++i) {
+	refstrprint(d->buckets[i], NULL);
+    }
 }
 #endif