+++ CHANGELOG.md
@@ -56,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: The `Agraph_t` structure has gained a new opaque field `g_seq2`.
 - **Breaking**: The `EXPANDBP` macro has been removed.
 - **Breaking**: The `EXPANDBB` macro has been removed.
+- **Breaking**: `Agclos_t.strdict` is now an opaque pointer.
 
 ### Fixed